Move SectionWithActiveInfo into the plot package and rename as Plot
