+++ packages/core/src/index.ts
@@ -1,7 +1,5 @@
 export * from './components/PageProvider';
 export * from './components/Section';
-export * from './section/SectionWithActiveInfo';
-export * from './section/StickySectionWithActiveInfo';
 export * from './context/PageContext';
 export * from './utils/getIntersectionObserver';
 export * from './utils/getStickyPosition';
