+++ packages/plot/src/index.ts
@@ -1,0 +1,3 @@
+export * from './hooks/useActiveSectionInfo';
+export * from './components/Plot';
+export * from './components/StickyPlot';
